requirements.txt
cx_Freeze
PyQt5
PyYaml